Add popup form to create a new book
I created a form which is hidden on page load that changes to the
active pseudostate once the "+ Add new book" button is clicked. The
form pop ups and can be cancelled with the "x" button in the top right
corner. The form has some simple styling and the layout is set with both
grid and flexbox, but there is still some styling to be done and the
submit button is not yet set up to add the book to the library once
clicked (it does nothing for the moment).