Add submit functionality to 'Add New Book' button
I added a function on click of the 'submit' button in the form that is
brought up once a user wishes to add a new book to the library. This
function takes in the values inputed by the user, creates a new book
object, then uses the addToLibrary() function to add the new book into
the myLibrary array. I also edited the displayCards() function to clear
the content each time it is run, because I had the error of duplicating
what was already in the library. I also created the clearFormValues()
function which is run when clicking the 'Add New Book' button so that
the form is empty each time a user wishes to add a new book.