Add remove button function and toggle read status
Was too excited after I got the remove button to work that I forgot to
do a commit so this commit is two parts. I added the functionality to
the remove button under each book object in the library so that it
deletes the associated object only. This is done within the displayCards()
function and takes the unique ID of the button when clicked and takes the
value of the button (which is the same as the associated book index) and
uses that value to splice the book from the array, finally running the
displayCards() function again to update the display.

Also added a toggleRead() function to the Book prototype which toggles
the boolean of this.isRead. I added this functionality to a button on
each book object using the same principles described above.